+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - added controller detection during runtime (#850)
 - added an option to allow cancelling Lara's ledge-swinging animation (#856)
 - added an option to allow Lara to jump at any point while running, similar to TR2+ (#157)
+- added the ability to define the anchor room for Bacon Lara in the gameflow (#868)
 - changed screen resolution option to apply immediately (#114)
 - changed shaders to use GLSL 1.20 which should fix most issues with OpenGL 2.1 (#327, #685)
 - fixed sounds stopping instead of pausing if game sounds in inventory are disabled (#717)